+++ src/library/adwr/CMakeLists.txt
