+++ src/library/adwr/binary.cpp
