+++ src/library/adwr/def.hpp
