+++ src/library/adwr/math.cpp
