+++ src/library/adwr/math.hpp
