+++ src/library/adwr/stream.cpp
