+++ src/library/adwr/stream.hpp
