+++ src/library/adwr/string.cpp
