+++ src/library/adwr/string.hpp
